Fix local development database name setting
In v0.3.0 the Docker Compose database setup was adjusted, part of which
was switching the `POSTGRES_USER` setting from `app_rails` to `app`,
aligning with the other templates. The PostgreSQL Docker image uses the
value of `POSTGRES_USER` as the database name as well (if the database
name isn't specified explicitly).

So need to adjust the application's settings to match.