⚠️ Remove create user roles migration (#83)
The commit
https://github.com/navapbc/template-application-rails/commit/5e2a120c1999c5b3632aac800526d6bfc862a734
removes the user roles from the application code but not from the
database. This change removes it from migrations. This change was
separated out since removing an already existing migration is not
backwards compatible. Since this is a template it makes sense to do it
this way rather than add an extra migration that removes the user_roles
table.

## Migration notes

When updating the template in this version, revert the changes in these
files:
- <APP_NAME>/db/migrate/20240410213056_create_user_roles.rb (deleted in
this change)
- <APP_NAME>/db/schema.rb (user_tables removed)