Force RAILS_ENV=test in RSpec config (#117)
It's not clear to me why you'd ever run the test suite against a
different `RAILS_ENV` setting. And even if you did want to do that, why
you'd want the test suite configuration for selecting that `RAILS_ENV`
to use be the same environment variable the real application uses, which
creates conflicts and confusion when you want to run both the
application and the test suite in the same host environment (i.e.,
during all of development).

The test suite should be as self contained as possible. And while not
the end of the world, shouldn't *have* to use a wrapper script for the
test suite to minimally function.