Resolve form_with deprecation warning
Passing `nil` to the `:model` argument in `form_with` has been
deprecated and will raise an exception in later versions of Rails.
https://github.com/rails/rails/pull/50931

This commit resolves the deprecation warning when a user attempts to
update their email address with one that is invalid. Previously, this
would re-render the edit view without setting @password_form causing the
warning to be displayed. Now, the Users::Accounts controller has a
`set_password_form` callback to ensure the password_form is set and the
edit view renders correctly.