Tweak language selector (#134)
Two main changes:

- Add some slight additional padding to the element on desktop to
vertically align better with other the header elements. Not the cleanest
solution, but simple with existing markup.
- Move the language selector out of the menu on mobile.

This better follows USWDS guidelines[1] around not combining with other
navigation items on mobile and implementations[2]. More accurate to the
guidelines would be to have the language selector outside the main nav
area altogether/not to have duplicate buttons in the markup, but that's
a larger change.

[1]
https://designsystem.digital.gov/patterns/select-a-language/two-languages/
[2] https://www.usa.gov/